add validations for inconsistency states for scanners
