+++ ydb/core/formats/arrow/reader/ya.make
@@ -9,6 +9,7 @@ PEERDIR(
 )
 
 SRCS(
+    batch_iterator.cpp
     merger.cpp
     position.cpp
     heap.cpp
